Add sass and styles to project
