+++ src/Components/Header/index.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+import Wrapper from "../Wrapper";
+
+const Header = () => (
+    <header className="header">
+        <Wrapper />
+    </header>
+);
+
+export default Header;
