+++ src/Components/Wrapper/index.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+const Wrapper = () => <div className="wrapper"></div>
+
+export default Wrapper;
