+++ src/Styles/sass/abstract/_constants.scss
@@ -1,0 +1,9 @@
+$color-dark: #000000;
+$color-light: #ffffff;
+
+$color-primary: #000;
+$color-secondary:#f4f7f6;
+
+$content-width: 1400px;
+$tablet-width: 960px;
+$mobile-width: 640px;
