+++ src/Styles/sass/base/_base.scss
@@ -1,0 +1,19 @@
+@import url('https://fonts.googleapis.com/css?family=Roboto');
+
+html {
+    font-family: 'Roboto', sans-serif;
+    overflow: hidden;
+}
+
+body {
+    margin: 0;
+    color: $color-dark;
+    display: flex;
+    flex-direction: column;
+    min-height: 100vh;
+    @include font($size: 10px);
+}
+
+h1 {
+    @include font($size: 2.4rem);
+}
