+++ src/Styles/sass/layout/_header.scss
@@ -1,0 +1,3 @@
+.header {
+    background: $color-primary;
+  }
