+++ src/Styles/sass/layout/_wrapper.scss
@@ -1,0 +1,4 @@
+.wrapper {
+    max-width: $content-width;
+    margin: 0 auto;
+  }
