+++ src/Styles/sass/style.scss
@@ -1,0 +1,10 @@
+//Abstract
+@import "abstract/constants";
+@import "abstract/mixins";
+
+// base
+@import './base/base';
+
+//Layouts
+@import "layout/header";
+@import "layout/wrapper";
