+++ src/index.js
@@ -1,5 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
+import './Styles/sass/style.scss';
 
 import App from "./App.jsx";
 const rootElement = document.getElementById("root");
